Loads of stuff with help, also added interpreter support

+++ OpenIDE.Core.Tests/Config/ConfigurationTests.cs
@@ -28,7 +28,8 @@ namespace OpenIDE.Core.Tests.Config
 		[TearDown]
 		public void Teardown()
 		{
-			File.Delete(_config.ConfigurationFile);
+			if (File.Exists(_config.ConfigurationFile))
+				File.Delete(_config.ConfigurationFile);
 		}
 
 		[Test]
@@ -77,6 +78,54 @@ namespace OpenIDE.Core.Tests.Config
 		{
 			_config.Write("enabled.languages-=C#");
 			configIs("");
+		}
+
+		[Test]
+		public void When_given_an_unexisting_setting_we_get_null_returned() 
+		{
+			var setting = _config.Get("mytest.setting");
+			Assert.That(setting, Is.Null);
+		}
+
+		[Test]
+		public void When_given_an_exact_setting_we_can_read_it() 
+		{
+			_config.Write("mytest.setting.with.more=20");
+			_config.Write("mytest.setting=15");
+			var setting = _config.Get("mytest.setting");
+			Assert.That(setting.Key, Is.EqualTo("mytest.setting"));
+			Assert.That(setting.Value, Is.EqualTo("15"));
+		}
+		
+		[Test]
+		public void When_given_settings_starting_with_a_token_we_can_fetch_them() 
+		{
+			_config.Write("mytest.setting.with.more=20");
+			_config.Write("		 mytest.setting=15");
+			var settings = _config.GetSettingsStartingWith("mytest.setting");
+			Assert.That(settings.Length, Is.EqualTo(2));
+			Assert.That(settings[0].Key, Is.EqualTo("mytest.setting.with.more"));
+			Assert.That(settings[0].Value, Is.EqualTo("20"));
+			Assert.That(settings[1].Key, Is.EqualTo("mytest.setting"));
+			Assert.That(settings[1].Value, Is.EqualTo("15"));
+		}
+
+		[Test]
+		public void When_splitting_values_of_a_setting_with_no_value_returns_an_empty_list()
+		{
+			Assert.That(
+				new ConfigurationSetting("key", null).SplitBy(",").Length,
+				Is.EqualTo(0));
+		}
+
+		[Test]
+		public void When_splitting_values_it_will_return_array_of_values()
+		{
+			var values = new ConfigurationSetting("key", "1,2,3").SplitBy(",");
+			Assert.That(values.Length, Is.EqualTo(3));
+			Assert.That(values[0], Is.EqualTo("1"));
+			Assert.That(values[1], Is.EqualTo("2"));
+			Assert.That(values[2], Is.EqualTo("3"));
 		}
 
 		private void configIs(string content)
